New backwards-incompatible cache layout
With the new layout, instead of adding two key-value pairs
(one for metadata, one for the blob) for each pair, we add
just one, using nippy’s `thaw-from-in!` and `freeze-to-out!`.
This has a benefit that we can now easily implement Seqable
on RocksDB caches and easily iterate over all keys, metas
and blobs.

Also, fix a bug where the cache would pass HeaderMaps through
nippy, which were then unthawable back.